⚡️ improvement and rename index.ts
