+++ src/services/index.ts
@@ -1,4 +1,6 @@
 export * from "./join";
+export * from "./member";
+export * from "./mention";
+export * from "./slashCommand";
 export * from "./vc";
 export * from "./webhook";
-export * from "./member";
